Conflicts:
	holiday/src/main/java/org/kernely/holiday/dto/HolidayDTO.java
	holiday/src/main/java/org/kernely/holiday/dto/HolidayUpdateRequestDTO.java
	holiday/src/main/java/org/kernely/holiday/model/HolidayRequestDetail.java
	holiday/src/main/java/org/kernely/holiday/service/HolidayService.java

+++ holiday/src/main/java/org/kernely/holiday/model/HolidayRequestDetail.java
@@ -51,7 +51,7 @@ public class HolidayRequestDetail extends AbstractModel {
     @JoinColumn(name = "fk_holiday_type")
 	private HolidayType type;
 
-	
+
 	/**
 	 * @return the id
 	 */
@@ -63,7 +63,7 @@ public class HolidayRequestDetail extends AbstractModel {
 	 * @param id
 	 *            the id to set
 	 */
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
